refactor(Gridbox, Listbox): extract `queryCheckedElement` function into `@/shared`
