--- packages/core/src/Gridbox/utils.ts
@@ -1,3 +1,0 @@
-export function queryCheckedElement(container: HTMLElement): HTMLElement | null {
-  return container?.querySelector('[data-state="checked"]') as HTMLElement | null
-}
