--- packages/core/src/Listbox/utils.ts
@@ -1,3 +1,0 @@
-export function queryCheckedElement(parentEl: HTMLElement | null) {
-  return parentEl?.querySelector('[data-state=checked]') as HTMLElement | null
-}
