+++ packages/core/src/shared/dom.ts
@@ -1,0 +1,3 @@
+export function queryCheckedElement(parentEl: HTMLElement | null) {
+  return parentEl?.querySelector('[data-state=checked]') as HTMLElement | null
+}
