+++ packages/core/src/shared/index.ts
@@ -3,6 +3,7 @@ export * from './browser'
 export * from './clamp'
 export { createContext } from './createContext'
 export { type DateRange, type DateStep, type DateValue } from './date'
+export * from './dom'
 export { getActiveElement } from './getActiveElement'
 export { handleAndDispatchCustomEvent } from './handleAndDispatchCustomEvent'
 export { isValidVNodeElement } from './isValidVNodeElement'
@@ -33,5 +34,5 @@ export { useSelectionBehavior } from './useSelectionBehavior'
 export { useSize } from './useSize'
 export { useStateMachine } from './useStateMachine'
 export { useTypeahead } from './useTypeahead'
-export { compare, valueComparator } from './value'
+export * from './value'
 export { withDefault } from './withDefault'
